Consolidate the venue-map REST surface into Venue\Map\Rest_Api
Rest_Combo held the combo arg schema and request parsing while route
registration and the regenerate handler stayed on Map, splitting the
subsystem's REST surface across two classes — one with a name that fit
neither the domain-noun convention (Manager, Prewarm, Dimensions) nor
the Rest_Api layer convention established by Event\Rest_Api. The new
Venue\Map\Rest_Api owns all of it: route registration, route_args(),
parse_request(), and rest_regenerate(). Map keeps the domain operations
(regenerate, ensure_combo, descriptors) and sheds more complexity;
parse_coord() becomes public so the handler reuses the same coercion.
Endpoint URL and behavior are unchanged.
