+++ test/unit/php/includes/tests/core/classes/venue/map/class-test-map.php
@@ -144,12 +144,6 @@ class Test_Map extends Base {
 			),
 			array(
 				'type'     => 'action',
-				'name'     => 'rest_api_init',
-				'priority' => 10,
-				'callback' => array( $instance, 'register_rest_routes' ),
-			),
-			array(
-				'type'     => 'action',
 				'name'     => 'update_option_gatherpress_settings',
 				'priority' => 10,
 				'callback' => array( $instance, 'maybe_handle_settings_change' ),
@@ -1472,16 +1466,9 @@ class Test_Map extends Base {
 	public function test_parse_coord(): void {
 		$instance = Map::get_instance();
 
-		$this->assertSame(
-			37.3318,
-			Utility::invoke_hidden_method( $instance, 'parse_coord', array( '37.3318' ) )
-		);
-		$this->assertNull(
-			Utility::invoke_hidden_method( $instance, 'parse_coord', array( 'not a number' ) )
-		);
-		$this->assertNull(
-			Utility::invoke_hidden_method( $instance, 'parse_coord', array( '' ) )
-		);
+		$this->assertSame( 37.3318, $instance->parse_coord( '37.3318' ) );
+		$this->assertNull( $instance->parse_coord( 'not a number' ) );
+		$this->assertNull( $instance->parse_coord( '' ) );
 	}
 
 	/**
@@ -2172,120 +2159,6 @@ class Test_Map extends Base {
 	}
 
 	/**
-	 * The POST /venue/{id}/regenerate-map endpoint requires edit_post on
-	 * the target venue — anonymous callers receive 401/403.
-	 *
-	 * @covers ::register_rest_routes
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_requires_edit_post(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$post_id = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		wp_set_current_user( 0 );
-
-		$request  = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$response = rest_do_request( $request );
-
-		$this->assertGreaterThanOrEqual( 400, $response->get_status() );
-		$this->assertLessThan( 500, $response->get_status() );
-	}
-
-	/**
-	 * Happy-path REST call: an editor-level user regenerates a venue with
-	 * usable coordinates and gets the fresh descriptor map in the response.
-	 *
-	 * @covers ::register_rest_routes
-	 * @covers ::rest_regenerate
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_returns_fresh_descriptors(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$editor_id = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$post_id   = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		add_post_meta( $post_id, 'gatherpress_address', '1 Infinite Loop' );
-		add_post_meta( $post_id, 'gatherpress_latitude', '37.3318' );
-		add_post_meta( $post_id, 'gatherpress_longitude', '-122.0312' );
-
-		wp_set_current_user( $editor_id );
-
-		$request  = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$response = rest_do_request( $request );
-
-		$this->assertSame( 200, $response->get_status() );
-
-		$data = $response->get_data();
-		$this->assertArrayHasKey( 'descriptors', $data );
-		$this->assertSame( '', $data['reason'] );
-
-		$default_key = sprintf(
-			'%dx%dx%dx%s',
-			Map::DEFAULT_ZOOM,
-			Map::DEFAULT_HEIGHT * 2,
-			Map::DEFAULT_HEIGHT,
-			Map::DEFAULT_MAP_TYPE
-		);
-		$this->assertArrayHasKey( 'osm', (array) $data['descriptors'] );
-		$this->assertArrayHasKey( $default_key, (array) $data['descriptors']['osm'] );
-	}
-
-	/**
-	 * Ensure_only lazily generates one combo without wiping other descriptors.
-	 *
-	 * @covers ::register_rest_routes
-	 * @covers ::rest_regenerate
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_ensure_only_generates_requested_combo(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$editor_id = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$post_id   = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		add_post_meta( $post_id, 'gatherpress_address', '1 Infinite Loop' );
-		add_post_meta( $post_id, 'gatherpress_latitude', '37.3318' );
-		add_post_meta( $post_id, 'gatherpress_longitude', '-122.0312' );
-
-		wp_set_current_user( $editor_id );
-
-		$request = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$request->set_param( 'zoom', 15 );
-		$request->set_param( 'width', 800 );
-		$request->set_param( 'height', 400 );
-		$request->set_param( 'map_type', 'hybrid' );
-		$request->set_param( 'ensure_only', true );
-
-		$response = rest_do_request( $request );
-
-		$this->assertSame( 200, $response->get_status() );
-
-		$data = $response->get_data();
-		$this->assertSame( '', $data['reason'] );
-		$this->assertArrayHasKey(
-			'15x800x400xhybrid',
-			(array) $data['descriptors']['osm']
-		);
-	}
-
-	/**
 	 * Direct ensure_combo() coverage: generates the requested combo and
 	 * returns the venue's full descriptor map.
 	 *
@@ -2334,218 +2207,6 @@ class Test_Map extends Base {
 			$instance->ensure_combo( $post_id, array( 'zoom' => 15 ) ),
 			'A venue without coordinates cannot produce the combo.'
 		);
-	}
-
-	/**
-	 * When the venue has no address, the REST endpoint returns a 200 with
-	 * a structured reason so the client can render the appropriate
-	 * placeholder rather than a generic error state.
-	 *
-	 * @covers ::rest_regenerate
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_reports_no_address_reason(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$editor_id = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$post_id   = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		wp_set_current_user( $editor_id );
-
-		$request  = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$response = rest_do_request( $request );
-
-		$this->assertSame( 200, $response->get_status() );
-		$data = $response->get_data();
-		$this->assertSame( 'no_address', $data['reason'] );
-	}
-
-	/**
-	 * When the address is set but not yet geocoded (empty lat/lng), the
-	 * endpoint reports `awaiting_geocode` so the client shows the "Save
-	 * the venue first" placeholder instead of treating it as a failure.
-	 *
-	 * @covers ::rest_regenerate
-	 *
-	 * @return void
-	 */
-	/**
-	 * Reports a `generation_failed` reason when the venue has coordinates
-	 * but every combo's PNG write fails. Simulated here by putting the
-	 * uploads dir in an error state so `save_image` returns null and the
-	 * regenerate() call comes back with an empty descriptor map.
-	 *
-	 * @covers ::rest_regenerate
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_reports_generation_failed_when_saves_fail(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$editor_id = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$post_id   = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		add_post_meta( $post_id, 'gatherpress_address', '1 Infinite Loop' );
-		add_post_meta( $post_id, 'gatherpress_latitude', '37.3318' );
-		add_post_meta( $post_id, 'gatherpress_longitude', '-122.0312' );
-
-		wp_set_current_user( $editor_id );
-
-		// Force save_image to fail for every combo so regenerate() returns
-		// an empty array and the REST handler enters the generation_failed
-		// branch.
-		$force_error = static function ( $dirs ) {
-			$dirs['error'] = 'Simulated uploads failure.';
-			return $dirs;
-		};
-		add_filter( 'upload_dir', $force_error );
-
-		$request = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$request->set_param( 'zoom', 15 );
-		$request->set_param( 'width', 800 );
-		$request->set_param( 'height', 400 );
-		$request->set_param( 'aspect_ratio', '2/1' );
-
-		$response = rest_do_request( $request );
-
-		remove_filter( 'upload_dir', $force_error );
-
-		$this->assertSame( 200, $response->get_status() );
-		$data = $response->get_data();
-		$this->assertSame( 'generation_failed', $data['reason'] );
-	}
-
-	/**
-	 * Rejects a malformed `aspect_ratio` parameter at the REST boundary
-	 * with a 400 instead of silently falling back. Empty / valid values
-	 * (slash or colon form) pass through.
-	 *
-	 * @covers ::register_rest_routes
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_aspect_ratio_validator(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$editor_id = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$post_id   = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		add_post_meta( $post_id, 'gatherpress_address', '1 Infinite Loop' );
-		add_post_meta( $post_id, 'gatherpress_latitude', '37.3318' );
-		add_post_meta( $post_id, 'gatherpress_longitude', '-122.0312' );
-
-		wp_set_current_user( $editor_id );
-
-		$invalid = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$invalid->set_param( 'aspect_ratio', 'not-a-ratio' );
-
-		$this->assertSame(
-			400,
-			rest_do_request( $invalid )->get_status(),
-			'A malformed aspect_ratio string is rejected at the REST boundary.'
-		);
-
-		$valid_slash = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$valid_slash->set_param( 'aspect_ratio', '16/9' );
-		$this->assertSame(
-			200,
-			rest_do_request( $valid_slash )->get_status(),
-			'Slash-form aspect ratio passes.'
-		);
-
-		$valid_colon = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$valid_colon->set_param( 'aspect_ratio', '4:3' );
-		$this->assertSame(
-			200,
-			rest_do_request( $valid_colon )->get_status(),
-			'Colon-form aspect ratio passes.'
-		);
-
-		$empty = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$empty->set_param( 'aspect_ratio', '' );
-		$this->assertSame(
-			200,
-			rest_do_request( $empty )->get_status(),
-			'Empty aspect ratio is treated as "use server default" and passes.'
-		);
-
-		// Degenerate `0/X` / `X/0` values that CSS would treat as auto
-		// must still be rejected — the block never emits a zero side.
-		$zero_numerator = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$zero_numerator->set_param( 'aspect_ratio', '0/9' );
-		$this->assertSame(
-			400,
-			rest_do_request( $zero_numerator )->get_status(),
-			'Zero numerator is rejected by the tightened [1-9] rule.'
-		);
-
-		$zero_denominator = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$zero_denominator->set_param( 'aspect_ratio', '9/0' );
-		$this->assertSame(
-			400,
-			rest_do_request( $zero_denominator )->get_status(),
-			'Zero denominator is rejected by the tightened [1-9] rule.'
-		);
-	}
-
-	/**
-	 * Reports the `awaiting_geocode` reason when the venue has an address
-	 * but no resolved coordinates yet.
-	 *
-	 * @covers ::rest_regenerate
-	 *
-	 * @return void
-	 */
-	public function test_rest_regenerate_reports_awaiting_geocode_reason(): void {
-		$instance = Map::get_instance();
-		$instance->register_rest_routes();
-
-		$editor_id = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$post_id   = $this->factory->post->create( array( 'post_type' => Venue::POST_TYPE ) );
-
-		add_post_meta( $post_id, 'gatherpress_address', 'Somewhere' );
-		add_post_meta( $post_id, 'gatherpress_latitude', '' );
-		add_post_meta( $post_id, 'gatherpress_longitude', '' );
-
-		wp_set_current_user( $editor_id );
-
-		$request  = new \WP_REST_Request(
-			'POST',
-			sprintf( '/%s/venue/%d/regenerate-map', GATHERPRESS_REST_NAMESPACE, $post_id )
-		);
-		$response = rest_do_request( $request );
-
-		$this->assertSame( 200, $response->get_status() );
-		$data = $response->get_data();
-		$this->assertSame( 'awaiting_geocode', $data['reason'] );
 	}
 
 	/**
